backend/agentic/requirements.txt
fastapi==0.120.4
pydantic==2.12.3
python-dotenv==1.1.1
uvicorn==0.38.0
psycopg2-binary==2.9.11
sqlalchemy==2.0.44
langchain==1.0.3
langchain_core==1.0.3
langchain_groq==1.0.0
langgraph==1.0.2
langchain_community==0.4.1
litellm==1.80.5
sqlalchemy==2.0.44